ignore if nothing to delete

+++ configuration/views.py
@@ -59,10 +59,11 @@ def handle_boto3_upload(f):
         for obj in objs:
             if obj["Key"] != new_key:
                 to_delete_keys.append({"Key": obj["Key"]})
-        s3 = boto3.resource('s3', aws_access_key_id=settings.AWS_ACCESS_KEY_ID,
-                            aws_secret_access_key=settings.AWS_SECRET_ACCESS_KEY)
 
-        s3.meta.client.delete_objects(Bucket=settings.AWS_STORAGE_BUCKET_NAME, Delete={"Objects": to_delete_keys})
+        if len(to_delete_keys) > 0:
+            s3 = boto3.resource('s3', aws_access_key_id=settings.AWS_ACCESS_KEY_ID,
+                                aws_secret_access_key=settings.AWS_SECRET_ACCESS_KEY)
+            s3.meta.client.delete_objects(Bucket=settings.AWS_STORAGE_BUCKET_NAME, Delete={"Objects": to_delete_keys})
 
 
 @login_required
